+++ tests/integration/helpers/test_scan.py
@@ -70,10 +70,11 @@ class TestScanHelper(BaseTest):
         for _ in range(5):
             scan.launch().wait_or_cancel_after(10)
             assert scan.stopped(), u'Scan is stopped.'
+        scan_histories = scan.histories()
         most_recent_history = scan.last_history()
-        assert max([h.history_id for h in scan.histories()]) == most_recent_history.history_id, \
+        assert max([h.history_id for h in scan_histories]) == most_recent_history.history_id, \
             u'last_history should return the most recent history.'
-        assert max([h.last_modification_date for h in scan.histories()]) == most_recent_history.last_modification_date, \
+        assert max([h.last_modification_date for h in scan_histories]) == most_recent_history.last_modification_date, \
             u'last_history should return the most recent history.'
 
     @pytest.mark.xfail(reason="CI-16090")